chore(): renaming and moving components
