+++ src/components/Properties/Property/styles.js
